#17 - JavaDoc and code polish.
Added a bit of explaining JavaDoc to VideoShop main class and the DataInitializer implementation.

Simplified code in BossController. Adapted to new API in Inventory in CatalogController and added a special flag to indicate the ability to order an item by checking its quantity. Added not-null assertions to ShopController.

Added a few i18n properties and used them in the navigation view. Added auto-focus to user name field in login view.