#35 - Improve test case setup for integration tests involving security.
To be able to execute integration tests with security involved we now include the spring-security-test artifact (in version 4.0.0.RC1) as it provides the necessary API to authenticate requests made through Spring Test MVC. 

Altered AbstractWebIntegrationTest to wire Spring Security's FilterChainProxy into the MockMvc instance to be built. We also populate the ServletContext as for whatever reason the MVC test infrastructure only registers the context for a servlet, not as root application context. See [0] for reference.

The additional API of the test module now allows us to use ….with(…) clauses to fake authentications to the request execution. See SecurityMockMvcRequestPostProcessors for details.

Not how we altered the test case for the not allowed access as the now in place infrastructure translates the original AuthenticationException into a redirect to the login page.

Upgraded to Salespoint 6.1.1.BUILD-SNAPSHOT to benefit from encoding fix already available there.

[0] https://jira.spring.io/browse/SPR-12553