#67 - Refactoring towards Bounded Contexts.
Re-organized code into core business domains to better align with the actual domain. We now have packages for:

- Catalog

Contains extends of Salespoint's Product abstraction to represent Discs (DVD, Blueray) and Comments. It contains the DataInitializer to create Discs at application start and a Controller to browse the catalog and show details about each disc allowing to create Comments.

- Customer

Contains the Customer abstraction alongside a repository for it. Customers consist of a reference to a Salespoint UserAccount and further data. It exposes API to create Customers via CustomerManagement and a Controller to register customers and show all customers in the system (admin access only).

- Inventory

Contains a DataInitializer to initialize a default stock of 10 items for each Disc registered in the Catalog. Exposes a Controller to list all InventoryItems currently available in the system.

- Order

Exposes a Controller to add Discs to the Salespoint Cart and eventually conclude the order. Also allows to see all orders (admin access only).

This restructuring allowed the following improvements to the overall system:

- Web controllers have less and more focussed dependencies. Previously, the structure of the controllers were centered around a mixture of business functionality and permissions (BossController). We now only separate by domain, so that e.g. the OrderController contains resources for both public and admin access.
- We can make more Java classes package protected and thus avoid them being used from invalid other places in the codebase.

Miscellaneous:

- DataInitializer implementations are split up, made package protected and ordered via Spring's @Order.
- CustomerManagement has now both UserAccount creation (via Salespoint's UserAccountManager) and Customer creation (via CustomerRepository) wrapped into a single transactional method, so that the code in CustomerController boils down to a single line of code.
- Polished URI space and the creation of links from within the Thymeleaf templates.