#119 - Consistent removal of ROLE_ prefixes in role checks.
We had a few leftovers of the usage of ROLE_ in security handling that are now consistently removed. This requires the upgrade to Salespoint 7.2.1 as it fixes the exposure of role names lacking the ROLE_ prefix to Spring security.

Also refactored CustomDataInitializer to use CustomerManagement instead of CustomerRepository to avoid having to replicate the functionality of CustomerManagement.createCustomer(…) as this already combines the creation of a UserAccount and a Customer in the first place.