#120 - Added Modulith component diagrams.
VideoshopModularityTests now contains a test method to create an overall component diagram as well as a module centric one for each of the business modules contained in Videoshop. They can be found in target/generated-docs/moduliths.

Upgrade to Modulith 1.0 RC3, as it exposes new options for coloring and display name customizations.