#117 - Reference UniqueInventory in UML models and docs.
With commit 848d0c5 the videoshop was adapted to the new Inventory API
(UniqueInventory/UniqueInventoryItem instead of Inventory/InventoryItem).

This commit updates the UML design model, the inventory class diagram
and the developer documentation to reference the new classes.

Note that we do *not* update the spec (Pflichtenheft), because in that
document the names Inventory and InventoryItem refer to *domain terms*.
They only happen to have the same name in the Salespoint framework.

Also, git-ignore MagicDraw backup files from now on.