#135 - Add server side validation for comment and rating.
Handle binding/validation errors of `CommandAndRating` in the POST
mapping using an `Errors` object. Adapt the template to display current
binding and validation errors, if any.

The rating field in `CommentAndRating` had to be changed from a
primitive to a reference type to make it possible to initialize it with
missing values.