Updated architecural diagrams in Plichtenheft
The existing top-level and context diagrams were not consistent
with the use case diagram. That often resulted in confusion
between students.

Updated the diagrams to be more consistent with the use case diagram
but left some connections out that could miss in this stage of
analysis (also to keep the diagrams simple).